src/python/requirements.txt
openai>=1.0.0
pandas
python-dotenv>=0.21.0
flask>=2.0.0
litellm